the functionality has been removed in utils and iomodule
