--- lesson_05/homework_03/game.h
@@ -1,7 +1,0 @@
-#include <iostream>
-#include <vector>
-
-#pragma once
-
-
-bool check_str_in_arr(const std::string &external_str, const std::vector<std::string> &arr);
