+++ lesson_05/homework_03/iomodule/io.h
@@ -1,0 +1,8 @@
+#include <string>
+
+#pragma once
+
+
+bool write_to_score_board(const std::string &user_name, const int attempts);
+
+bool show_score_board();
